main.py
#!/usr/bin/env python3
# -*- coding: utf-8 -*-
"""
LeadScraper LATAM - Main Application

This is the main entry point for the LeadScraper LATAM application.
It orchestrates the scraping process from multiple sources (Google Maps, Instagram,
public directories), processes the collected data, and uploads it to Google Sheets.
"""

import os
import logging
from dotenv import load_dotenv

# Configure logging
logging.basicConfig(
    level=logging.INFO,
    format='%(asctime)s - %(name)s - %(levelname)s - %(message)s',
    handlers=[
        logging.FileHandler("leadscraper.log"),
        logging.StreamHandler()
    ]
)
logger = logging.getLogger(__name__)

# Load environment variables
load_dotenv()

def main():
    """
    Main function to orchestrate the scraping, processing and uploading workflow.
    """
    logger.info("Starting LeadScraper LATAM")
    
    try:
        # TODO: Initialize scrapers
        
        # TODO: Run scraping process
        
        # TODO: Process and deduplicate data
        
        # TODO: Upload data to Google Sheets
        
        logger.info("LeadScraper LATAM completed successfully")
    except Exception as e:
        logger.error(f"An error occurred: {str(e)}", exc_info=True)
        return 1
    
    return 0

if __name__ == "__main__":
    exit_code = main()
    exit(exit_code)
