requirements.txt
# Web scraping
selenium==4.15.2
undetected-chromedriver==3.5.3
beautifulsoup4==4.12.2
requests==2.31.0

# Data processing
pandas==2.1.3
numpy==1.26.2

# Instagram scraping
instaloader==4.10.1

# Google Sheets integration
gspread==5.12.0
google-auth==2.23.4
google-auth-oauthlib==1.1.0

# Utils
python-dotenv==1.0.0
tqdm==4.66.1
pydantic==2.5.2

# For docs generation (optional, uncomment if needed)
# sphinx==7.2.6
# sphinx-rtd-theme==1.3.0
